wip: factor out matchers
